+++ pkg/deploy/kong_plugin.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 
 	s "github.com/vertigobr/safira/pkg/stack"
+	"github.com/vertigobr/safira/pkg/utils"
 )
 
 func (k *K8sYaml) MountKongPlugin(functionName, pluginName, namespace, env string) error {
@@ -16,19 +17,28 @@ func (k *K8sYaml) MountKongPlugin(functionName, pluginName, namespace, env strin
 
 	metadataName := fmt.Sprintf("%s-%s", functionName, pluginName)
 
+	repoName, err := utils.GetCurrentFolder()
+	if err != nil {
+		return err
+	}
+
 	*k = K8sYaml{
 		ApiVersion: "configuration.konghq.com/v1",
 		Kind:       "KongPlugin",
 		Metadata: metadata{
-			Name: metadataName,
+			Name:      metadataName,
 			Namespace: namespace,
 			Labels: map[string]string{
 				"global": stack.Functions[functionName].Plugins[pluginName].Global,
 			},
+			Annotations: map[string]string{
+				"safira.io/repository": repoName,
+				"safira.io/function":   functionName,
+			},
 		},
-		Config: stack.Functions[functionName].Plugins[pluginName].Config,
+		Config:     stack.Functions[functionName].Plugins[pluginName].Config,
 		ConfigFrom: stack.Functions[functionName].Plugins[pluginName].ConfigFrom,
-		Plugin: pluginName,
+		Plugin:     pluginName,
 	}
 
 	return nil