+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Features:
 Improvements:
 
 - Flag added to remove the function folder from the `safira function remove` command
+- Added new annotations in function deploy
+
+Bug Fixes:
+
+- Fixed error when hostname port was not declared in stack.yaml
 
 ## Version v0.0.9 - 2020-07-17